Update jody_hash, update copyright years, add CFLAGS_EXTRA

+++ jody_hash.c
@@ -5,7 +5,7 @@
  * a secure hash algorithm, but the calculation is drastically simpler
  * and faster.
  *
- * Copyright (C) 2014-2015 by Jody Bruchon <jody@jodybruchon.com>
+ * Copyright (C) 2014-2016 by Jody Bruchon <jody@jodybruchon.com>
  * Released under The MIT License or GNU GPL v2 (your choice)
  */
 
@@ -48,11 +48,11 @@ static const hash_t tail_mask[] = {
  * hash_t, it is MANDATORY that the caller provide a data buffer
  * which is divisible by sizeof(hash_t). */
 extern hash_t jody_block_hash(const hash_t * restrict data,
-		const hash_t start_hash, const unsigned int count)
+		const hash_t start_hash, const size_t count)
 {
 	register hash_t hash = start_hash;
 	register hash_t element;
-	unsigned int len;
+	size_t len;
 
 	/* Don't bother trying to hash a zero-length block */
 	if (count == 0) return hash;
