+++ jody_hash.h
@@ -1,6 +1,6 @@
 /* Jody Bruchon's fast hashing function (headers)
  *
- * Copyright (C) 2014-2015 by Jody Bruchon <jody@jodybruchon.com>
+ * Copyright (C) 2014-2016 by Jody Bruchon <jody@jodybruchon.com>
  * See jody_hash.c for more information.
  */
 
@@ -22,7 +22,7 @@ typedef uint64_t hash_t;
 #define JODY_HASH_VERSION 3
 
 extern hash_t jody_block_hash(const hash_t * restrict data,
-		const hash_t start_hash, const unsigned int count);
+		const hash_t start_hash, const size_t count);
 
 #ifdef __cplusplus
 }
